more [leagueId].tsx cleanup
create ScoreboardFooter component,
move BetInfo to separate file
